+++ csmed/datasets/csmed_cochrane/__init__.py
@@ -1,0 +1,2 @@
+from .csmed_cochrane import CSMeDCochrane
+
