--- csmed/datasets/datasets/csmed_cochrane/__init__.py
@@ -1,39 +1,0 @@
-TRAIN_REVIEWS = ['CD010206', 'CD011146', 'CD011721', 'CD009771', 'CD006962', 'CD007470', 'CD009946', 'CD006569',
-                 'CD006900', 'CD010910', 'CD007953', 'CD003160', 'CD011209', 'CD007505', 'CD006014', 'CD003559',
-                 'CD004920', 'CD011066', 'CD009436', 'CD004178', 'CD007736', 'CD011837', 'CD003353', 'CD001457',
-                 'CD009634', 'CD009633', 'CD011241', 'CD005019', 'CD002832', 'CD004903', 'CD003914', 'CD002062',
-                 'CD002250', 'CD009210', 'CD010333', 'CD009887', 'CD006832', 'CD008963', 'CD005614', 'CD010767',
-                 'CD006748', 'CD008873', 'CD006770', 'CD008482', 'CD001957', 'CD006172', 'CD010473', 'CD002783',
-                 'CD007037', 'CD008881', 'CD002117', 'CD009593', 'CD000284', 'CD002143', 'CD003266', 'CD011145',
-                 'CD008265', 'CD009185', 'CD010269', 'CD004250', 'CD005340', 'CD008807', 'CD004054', 'CD009125',
-                 'CD010266', 'CD008435', 'CD008056', 'CD010693', 'CD010235', 'CD000313', 'CD009403', 'CD008345',
-                 'CD002283', 'CD005260', 'CD001261', 'CD006612', 'CD010542', 'CD002073', 'CD007356', 'CD001431',
-                 'CD004520', 'CD009831', 'CD004376', 'CD010912', 'CD007719', 'CD004396', 'CD007525', 'CD010901',
-                 'CD012004', 'CD007115', 'CD010709', 'CD003987', 'CD009823', 'CD001843', 'CD002201', 'CD008969',
-                 'CD007122', 'CD011281', 'CD010533', 'CD010534', 'CD010479', 'CD008812', 'CD003855', 'CD008020',
-                 'CD007699', 'CD006546', 'CD006745', 'CD008213', 'CD004875', 'CD010845', 'CD009530', 'CD009537',
-                 'CD008081', 'CD012567', 'CD010239', 'CD012551', 'CD010864', 'CD011380', 'CD008054', 'CD008892',
-                 'CD012930', 'CD009323', 'CD011787', 'CD009519', 'CD010296', 'CD011975', 'CD012083', 'CD010386',
-                 'CD005253', 'CD009694', 'CD012281', 'CD012455', 'CD009642', 'CD011420', 'CD010705', 'CD012233',
-                 'CD010276', 'CD012521', 'CD008874', 'CD006715', 'CD010019', 'CD010213', 'CD007868', 'CD008686',
-                 'CD012347', 'CD011768', 'CD009551', 'CD010438', 'CD012165', 'CD011977', 'CD009372', 'CD010409',
-                 'CD009944', 'CD009579', 'CD012164', 'CD009786', 'CD007427', 'CD011126', 'CD012080', 'CD011053',
-                 'CD010526', 'CD011436', 'CD011431', 'CD010173', 'CD010778', 'CD007394', 'CD012010', 'CD011686',
-                 'CD010753', 'CD009044', 'CD006468', 'CD010355', 'CD008759', 'CD009647', 'CD010558', 'CD000996',
-                 'CD009020', 'CD010502', 'CD010680', 'CD011558', 'CD008018', 'CD009591', 'CD004414', 'CD012179',
-                 'CD010023', 'CD004069', 'CD008127', 'CD010847', 'CD010089', 'CD007020', 'CD006902', 'CD005128',
-                 'CD007428', 'CD002733', 'CD005083']
-
-EVAL_REVIEWS = ['CD005055', 'CD010038', 'CD002069', 'CD011912', 'CD009782', 'CD006127', 'CD011447', 'CD004241',
-                'CD009925', 'CD011602', 'CD010226', 'CD011724', 'CD010411', 'CD001298', 'CD011549', 'CD003557',
-                'CD006342', 'CD007497', 'CD011926', 'CD010375', 'CD006839', 'CD008426', 'CD000031', 'CD010387',
-                'CD008170', 'CD010657', 'CD010264', 'CD011047', 'CD010447', 'CD007379', 'CD007271', 'CD003137',
-                'CD012223', 'CD004288', 'CD000384', 'CD001150', 'CD002898', 'CD011515', 'CD007867', 'CD012599',
-                'CD005139', 'CD003504', 'CD008587', 'CD007028', 'CD000009', 'CD001800', 'CD012009', 'CD003067',
-                'CD004479', 'CD010632', 'CD010381', 'CD010425', 'CD006638', 'CD005346', 'CD009400', 'CD002764',
-                'CD009069', 'CD011732', 'CD009784', 'CD002115', 'CD008760', 'CD010339', 'CD010225', 'CD009263',
-                'CD012661', 'CD009669', 'CD011472', 'CD010633', 'CD008366', 'CD010927', 'CD008803', 'CD005397',
-                'CD010139', 'CD012342', 'CD009685', 'CD002840', 'CD012019', 'CD003344', 'CD012768', 'CD009135',
-                'CD007716', 'CD010685', 'CD005059', 'CD008185', 'CD006678', 'CD012069', 'CD011571', 'CD011140',
-                'CD011134', 'CD012669', 'CD007103', 'CD000014', 'CD012120', 'CD006995', 'CD010653', 'CD012216',
-                'CD009780', 'CD004944', 'CD007745', 'CD008392']
