+++ csmed/datasets/tar2018/tar2018.py
@@ -18,10 +18,10 @@ from typing import List, Tuple, Dict
 import datasets
 import pandas as pd
 
-from csmed.datasets.loader.bigbiohub import BigBioConfig
-from csmed.datasets.loader.bigbiohub import Tasks
-from csmed.datasets.loader.bigbiohub import text_features
-from csmed.datasets.utils import (
+from csmed.loader.bigbiohub import BigBioConfig
+from csmed.loader.bigbiohub import Tasks
+from csmed.loader.bigbiohub import text_features
+from csmed.utils import (
     is_prepared,
     get_from_pubmed,
     save_checksum,
