+++ visualisation/pages/05_✔️_Dataset_cards.py
@@ -9,7 +9,7 @@ from matplotlib import pyplot as plt
 from matplotlib_venn import venn2, venn3
 from rich import print as rprint
 
-from csmed.datasets.dataloader import BigBioConfigHelpers
+from csmed.dataloader import BigBioConfigHelpers
 from visualisation.ngram import get_tuples_manual_sentences
